Add hot water sensor
Big refactor, to keep api-instance for all sensors set up for the integration.
WIP.
