+++ custom_components/priceanalyzer/manifest.json
@@ -9,5 +9,5 @@
   "codeowners": ["@hellowlol", "@erlendsellie"],
   "iot_class": "cloud_polling",
   "requirements": ["nordpool>=0.2"],
-  "version": "0.0.10"
+  "version": "0.0.50"
 }
